requirements.txt
# VAPT-AI Dependencies
# ====================

# Core AI Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM Providers
langchain-nvidia-ai-endpoints>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# CLI Dashboard
rich>=13.0.0
textual>=0.80.0

# HTTP & Networking
httpx>=0.27.0
aiohttp>=3.10.0
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
asyncio-extra>=1.0.0

# Output parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0